backend/requirements.txt
fastapi
uvicorn
tensorflow-cpu
pillow
numpy
pydantic
python-multipart
scipy
